server_b/requirements.txt
Django>=5.0,<6.0
psycopg[binary]
djangorestframework
pika
drf-spectacular
django-environ
pytest
pytest-django
requests
PyJWT
